requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
defusedxml==0.7.1
Deprecated==1.2.14
diff-match-patch==20230430
Django==5.1.1
django-auditlog==3.0.0
django-bootstrap-icons==0.9.0
django-crispy-forms==2.3
django-extensions==3.2.3
django-import-export==4.1.1
django-jquery==3.1.0
django-tables2==2.7.0
djangorestframework==3.15.2
et-xmlfile==1.1.0
graphviz==0.20.3
gunicorn==23.0.0
idna==3.9
jaconv==0.4.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.1
pydot==3.0.1
pykakasi==2.3.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-environ==0.4.54
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
urllib3==2.2.3
wrapt==1.16.0
xlrd==2.0.1
xlwt==1.3.0
